pyproject.toml
[project]
name = "photo-scaler"
version = "0.1.0"
description = "Resizes images to a target maximum dimension and saves as optimized JPEG."
requires-python = ">=3.12"
dependencies = [
    "typer>=0.15.0",
    "rich>=14.3.3",
    "pillow>=10.2.0",
    "local-first-common",
    "toml>=0.10.2",
]

[project.scripts]
scale-photo = "photo_scaler.logic:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/photo_scaler"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
]

[tool.uv.sources]
local-first-common = {path = "../local-first-common", editable = true}

[tool.pytest.ini_options]
pythonpath = ["src"]
